Feature: UI draft
This branch introduces the simple UI layer. The main purpose is to establish the way of communication between the server module and the UI module, and thus set the foundation for further development. The current version of UI contains a number of text fields to input data manually. After CBOR implementation, the number of text fields will be reduced to one, as one serialized CBOR will contain all request/response data. Ultimately, the text fields will be replaced with audio transmission feature.

List of changes:
- created main_page_wrapper.dart and widgets which replace the previous command-line features with a graphic UI
- created main_page_cubit.dart to manage the main page states. Added a listener of the TrezorCommunicationNotifier class' variable, which affects certain states of the cubit.
- modified trezor_communication_notifier.dart to actually implement the notification feature - it informs about TrezorInteractiveRequest being processed and user interaction being required
- created trezor_event.dart as main object used for communication between server module and app module
- modified a_trezor_interactive_request.dart to let the subclasses provide the necessary data for the current version of UI
- reduced the necessary Trezor public key requests by saving the first requested public key in the external file and implementing a key derivation feature in pubkey_service.dart. The public key data storage will also be necessary for CBOR implementation.