Feature: Wallets preview
The purpose of this branch was to change the way the public key is displayed. Instead of showing a simplified preview intended mainly for development purposes, the goal was to make the display more informative for the user - to show which wallets can be accessed using the given key. To achieve this in a simple way, it was decided to display just the first few wallets (in the same format as WalletListItem on Snggle) as a preview. This way, the user can quickly confirm that the application has loaded the correct key, consistent with their Snggle wallets.

List of changes:
- removed the active pubkey from main_page.dart
- modified idle_page.dart - replaced "pubkeyExistsBool" with actual PubkeyModel and added "_deriveWallets()" method in to derive first 6 wallets of the uploaded pubkey
- created wallet_model.dart to represent wallet data on UI
- created wallet_list_preview.dart as a main widget for displaying the wallets and a series of its component widgets, that resemble Snggle UI in appearance
- moved cancel button from main_page.dart to data_transfer_page.dart and slightly corrected the padding in data_transfer_page.dart and manual_pubkey_upload.dart to make it more consistent